clean up old gallery code
