+++ modules/gallery/src/commonMain/kotlin/org/jetbrains/compose/storytale/gallery/Gallery.kt
@@ -3,6 +3,7 @@ package org.jetbrains.compose.storytale.gallery
 import androidx.compose.runtime.Composable
 import org.jetbrains.compose.storytale.gallery.material3.StorytaleGalleryApp
 
+@Suppress("unused") // The call of this function is added implicitly in the generated code
 @Composable
 fun Gallery() {
     StorytaleGalleryApp()
