--- modules/gallery/src/commonMain/kotlin/org/jetbrains/compose/storytale/gallery/compose/Remember.kt
@@ -1,11 +1,0 @@
-package org.jetbrains.compose.storytale.gallery.compose
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisallowComposableCalls
-import androidx.compose.runtime.remember
-
-@Composable
-inline fun <T, K> remembering(
-    key1: K,
-    crossinline calculation: @DisallowComposableCalls (K) -> T,
-): T = remember(key1) { calculation(key1) }
