--- modules/gallery/src/commonMain/kotlin/org/jetbrains/compose/storytale/gallery/navigation/Route.kt
@@ -1,15 +1,0 @@
-package org.jetbrains.compose.storytale.gallery.navigation
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data object Route {
-    @Serializable
-    data object MobileHome
-
-    @Serializable
-    data class MobileGallery(val storyId: Int)
-
-    @Serializable
-    data class DesktopGallery(val storyId: Int?)
-}
