+++ modules/gallery/src/commonMain/kotlin/org/jetbrains/compose/storytale/gallery/story/parameters/ListParameter.kt
@@ -4,9 +4,12 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.displayCutout
 import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.statusBarsPadding
@@ -30,9 +33,15 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.geometry.CornerRadius
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.storytale.gallery.generated.resources.Res
@@ -41,7 +50,6 @@ import org.jetbrains.compose.storytale.gallery.material3.ParameterDescription
 import org.jetbrains.compose.storytale.gallery.material3.ParameterHeader
 import org.jetbrains.compose.storytale.gallery.ui.component.CenterRow
 import org.jetbrains.compose.storytale.gallery.ui.component.Gap
-import org.jetbrains.compose.storytale.gallery.ui.component.sheet.StoryBottomSheetDragHandle
 
 @Suppress("ktlint:compose:mutable-state-param-check")
 @Composable
@@ -120,7 +128,7 @@ private fun <T> SelectionSheet(
     ModalBottomSheet(
         sheetState = sheetState,
         onDismissRequest = onDismissRequest,
-        dragHandle = { StoryBottomSheetDragHandle() },
+        dragHandle = { BottomSheetDragHandle() },
         modifier = Modifier
             .statusBarsPadding()
             .padding(top = (24 * 2).dp),
@@ -159,3 +167,30 @@ private fun <T> SelectionSheet(
         }
     }
 }
+
+
+@Composable
+private fun BottomSheetDragHandle(
+    modifier: Modifier = Modifier,
+    color: Color = Color(0xFFD2D7E6),
+) = Spacer(
+    modifier = modifier
+        .drawBehind {
+            val barWidthPx = barWidth.toPx()
+            val barHeightPx = barHeight.toPx()
+            val x = size.width / 2 - barWidthPx / 2
+            val y = size.height / 2 - barHeightPx / 2
+            drawRoundRect(
+                color = color,
+                topLeft = Offset(x, y),
+                size = Size(barWidthPx, barHeightPx),
+                cornerRadius = CornerRadius(barHeightPx / 2),
+            )
+        }
+        .fillMaxWidth()
+        .height(height),
+)
+
+private val height: Dp = 24.dp
+private val barWidth: Dp = 32.dp
+private val barHeight: Dp = 4.dp
