--- modules/gallery/src/commonMain/kotlin/org/jetbrains/compose/storytale/gallery/utils/ScreenSize.kt
@@ -1,15 +1,0 @@
-package org.jetbrains.compose.storytale.gallery.utils
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-
-expect object ScreenSize {
-    @get:Composable
-    val width: Dp
-
-    @get:Composable
-    val height: Dp
-}
-
-val ScreenSize.isMobile @Composable inline get() = width <= 480.dp
